Use ulong for lock idents consistently.

+++ source/include/esc/arch/eco32/thread.h
@@ -23,9 +23,9 @@
 
 static inline void locku(tULock *l) {
 	/* TODO eco32 has no atomic compare and swap instruction or similar :/ */
-	lock((uint)l,LOCK_EXCLUSIVE);
+	lock((ulong)l,LOCK_EXCLUSIVE);
 }
 
 static inline void unlocku(tULock *l) {
-	unlock((uint)l);
+	unlock((ulong)l);
 }
